# docs/openapi.yaml
openapi: 3.0.0
info:
  title: Bangalore Hospital Appointment API
  version: 1.0.0
  description: >
    Multi-hospital doctor appointment API (Bangalore, IST display / UTC storage).
    AI helpers are powered by OpenRouter (see /api/ai/*). Auth: JWT access
    (15m, Bearer) + rotating refresh in httpOnly cookie.
servers:
  - url: http://localhost:4000/api
security:
  - bearerAuth: []
...
paths:
...
  /hospitals:
    get:
      tags: [Hospitals]
      security: []
      summary: List hospitals (q, accreditation, emergency, type, pagination)
      parameters:
        - { name: q, in: query, schema: { type: string } }
        - { name: accreditation, in: query, schema: { type: string } }
        - { name: emergency, in: query, schema: { type: string, enum: ['true', 'false'] } }
        - { name: type, in: query, schema: { type: string, enum: [PRIVATE, GOVERNMENT] } }
        - { name: page, in: query, schema: { type: integer } }
        - { name: pageSize, in: query, schema: { type: integer } }
      responses:
        '200':
          description: Hospital list
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: '#/components/schemas/Hospital' } }
    post:
      tags: [Hospitals]
      summary: Create hospital (ADMIN)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HospitalInput' }
      responses:
        '201': { description: Created }
  /hospitals/{id}:
    get:
      tags: [Hospitals]
      security: []
      summary: Hospital detail
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
        '404': { description: Not found }
    put:
      tags: [Hospitals]
      summary: Update hospital (ADMIN)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/HospitalInput' }
      responses:
        '200': { description: OK }
    delete:
      tags: [Hospitals]
      summary: Delete hospital (ADMIN)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
      responses:
        '200': { description: OK }
...
  /patients/me:
    get:
      tags: [Patients]
      summary: Own patient profile (medical notes decrypted)
      responses:
        '200': { description: OK }
    put:
      tags: [Patients]
      summary: Update own profile (notes encrypted with AES-256-GCM)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string }
                phone: { type: string }
                dateOfBirth: { type: string, format: date-time, nullable: true }
                gender: { type: string, nullable: true }
                medicalNotes: { type: string, nullable: true }
      responses:
        '200': { description: OK }
  /patients/{id}:
    get:
      tags: [Patients]
      summary: Patient detail (ADMIN, audited, masked by default)
      parameters:
        - { name: id, in: path, required: true, schema: { type: string } }
        - { name: masked, in: query, schema: { type: string, default: 'true' } }
      responses:
        '200': { description: OK }
...
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Hospital:
      type: object
      properties:
        id: { type: string, example: '1' }
        slug: { type: string, example: apollo-hospital-bannerghatta-road }
        name: { type: string, example: 'Apollo Hospital, Bannerghatta Road' }
        address: { type: string, example: '154/11 Bannerghatta Main Road, Bangalore, Karnataka 560076, India' }
        contact: { type: string, example: '+91-80-68250000' }
        accreditation: { type: array, items: { type: string }, example: [NABH, JCI] }
        hasEmergency: { type: boolean }
        beds: { type: integer, nullable: true }
        dataVerified: { type: boolean }
    HospitalInput:
      type: object
      required: [name, type]
      properties:
        name: { type: string }
        type: { type: string, enum: [PRIVATE, GOVERNMENT] }
        address: { type: string, nullable: true }
        latitude: { type: number, nullable: true }
        longitude: { type: number, nullable: true }
        contact: { type: string, nullable: true }
        website: { type: string, nullable: true }
        accreditation: { type: array, items: { type: string } }
        hasEmergency: { type: boolean }
        beds: { type: integer, nullable: true }
        bedsNote: { type: string, nullable: true }
        keySpecialties: { type: array, items: { type: string } }
        dataVerified: { type: boolean }
        departments: { type: array, items: { type: string } }
...